pyproject.toml
[project]
name = "price-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.1",
    "fastapi>=0.121.0",
    "greenlet>=3.2.4",
    "pandas>=2.3.3",
    "psycopg[async,binary]>=3.2.12",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.4.10",
]
